\name{labplot2}
\alias{labplot2}
\title{
Scatterplot with labelled points
}
\description{
PLot vectors of points x and y, with each point labelled by values of a third vector.
}
\usage{
labplot2(x, y, labels, main = "", xlab = "", ylab = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Vector of x values
}
  \item{y}{
Vector of matching y values
}
  \item{labels}{
Vector of matching label values or codes
}
  \item{main}{
Main heading
}
  \item{xlab}{
X axis label
}
  \item{ylab}{
Y axis label
}
}
\value{
No value returned. Function is for its plotting side effect.
}
\author{
Neville Jackson
}
\examples{
}
\keyword{misc }
