\name{mfbw.df}
\alias{mfbw.df}
\docType{data}
\title{
Sheep foetal body weight data
}
\description{
Foetal weights in grams for Merino lambs from 53 to 137 days of gestation. 
}
\usage{data(mfbw.df)}
\format{
  A data frame with 43 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{a numeric vector - age in days}
    \item{\code{FoetWt}}{a numeric vector - weights in grams}
  }
}
\source{
Metcalf etal (1962) Quart. J. exp. Physiol. XLVII(4) 305-313
}
\examples{
data(mfbw.df)
}
\keyword{datasets}
