\name{mlbw.df}
\alias{mlbw.df}
\docType{data}
\title{
Merino lamb body weights from birth to 140 days
}
\description{
Merino lamb body weights for 4 groups of Merino lambs on different levels of nutrition
}
\usage{data(mlbw.df)}
\format{
  A data frame with 56 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{a numeric vector - age in days starting from birth}
    \item{\code{Group}}{a factor with levels \code{HH} \code{HL} \code{LH} \code{LL}}
    \item{\code{Weight}}{a numeric vector - weights of lambs in Kg}
  }
}
\source{
Taplan and Everitt(1962) URL https://pdfs.semanticscholar.org/90b6/db6e9763d25add182220c0b369023072d41b.pdf
}
\examples{
data(mlbw.df)
}
\keyword{datasets}
