\name{ppcell-internal}
\alias{varyonepar}
\alias{varytwopar}
\title{
Functions internal to ppcell
}
\description{
Internal functions not documented and not visible to user.
}
\usage{
varyonepar(param,which="cellbirthprob",values=seq(.0740,.0769,.0001))
varytwopar(param,which=c("pstarttime","sostarttime"),values=seq(63,65,.1))
}
\author{
Neville Jackson
}
\keyword{ internal }
