\name{ppcell-package}
\alias{ppcell-package}
\alias{ppcell}
\docType{package}
\title{
Calculate dynamics of pre-papilla cell numbers in developing sheep foetus
}
\description{
Calculations for investigating the dynaics of pre-papilla cell numbers in the developing sheep foetus, and its effect on follicle development.
}
\details{
\tabular{ll}{
Package: \tab ppcell\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2018-05-14\cr
License: \tab GPL-2 \cr
}
This package provides tools for simulation of the effect of pre-papilla cell numbers on follicle development, including plotting of results and investigating combinations of parameters which lead to particular outcomes.

You may wish to use this software if you are interested in wool follicle development and wool growth.

The main functions in this package are:
\describe{
\item{folli}{calculates cell numbers and follicle formation}
\item{fregion}{investigates feasable regions in the parameter space}
\item{plotxxx}{plotting routines}
}

There are also some datasets which are starting values for the 20 parameters required.
}
\author{
Neville Jackson

Maintainer: Neville Jackson <nanddjackson@bigpond.com>
}
\references{
  ppcell.pdf
}
\keyword{ package }
\examples{
}
