\name{ppcell-package}
\alias{ppcell-package}
\alias{ppcell}
\docType{package}
\title{
\packageTitle{ppcell}
}
\description{
\packageDescription{ppcell}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ppcell}
\packageIndices{ppcell}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ppcell}

Maintainer: \packageMaintainer{ppcell}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
